requirements.txt
ibm-cos-sdk>=2.13
pymilvus>=2.4
python-dotenv>=1.0
requests>=2.32
